+++ packages/ui/decorators/withTransition.ts
@@ -32,6 +32,11 @@ export interface TransitionInterface extends BaseInterface {
    */
   get targets(): HTMLElement[];
   /**
+   * Current state.
+   * @internal
+   */
+  state: null | 'entering' | 'leaving';
+  /**
    * Trigger the enter transition.
    */
   enter(target?: HTMLElement | HTMLElement[]): Promise<void>;
@@ -39,23 +44,54 @@ export interface TransitionInterface extends BaseInterface {
    * Trigger the leave transition.
    */
   leave(target?: HTMLElement | HTMLElement[]): Promise<void>;
+  /**
+   * Toggle the leave or enter transition.
+   * Defaults to the enter transition if no transition has been triggered yet.
+   */
+  toggle(target?: HTMLElement | HTMLElement[]): Promise<void>;
 }
 
 /**
  * Extend a class to add transition capabilities.
+ * @link https://ui.studiometa.dev/components/Transition/
  */
 export function withTransition<S extends Base>(
   BaseClass: typeof Base,
 ): BaseDecorator<TransitionInterface, S, TransitionProps> {
+  type TransitionConstructor<T extends Transition = Transition> = {
+    new (...args: any[]): T;
+    prototype: Transition;
+  } & Pick<typeof Transition, keyof typeof Transition>;
+
   /**
    * Class.
    */
   class Transition<T extends BaseProps = BaseProps> extends BaseClass<T & TransitionProps> {
     /**
+     * Declare the `this.constructor` type
+     * @link https://github.com/microsoft/TypeScript/issues/3841#issuecomment-2381594311
+     */
+    declare ['constructor']: TransitionConstructor;
+
+    /**
+     * Events.
+     */
+    static EVENTS = {
+      TRANSITION_TOGGLE: 'transition-toggle',
+      TRANSITION_ENTER: 'transition-enter',
+      TRANSITION_ENTER_START: 'transition-enter-start',
+      TRANSITION_ENTER_END: 'transition-enter-end',
+      TRANSITION_LEAVE: 'transition-leave',
+      TRANSITION_LEAVE_START: 'transition-leave-start',
+      TRANSITION_LEAVE_END: 'transition-leave-end',
+    } as const;
+
+    /**
      * Config.
      */
     static config: BaseConfig = {
       name: 'Transition',
+      emits: Object.values(this.EVENTS),
       options: {
         enterFrom: String,
         enterActive: String,
@@ -68,6 +104,19 @@ export function withTransition<S extends Base>(
         group: String,
       },
     };
+
+    /**
+     * States.
+     */
+    static STATES = {
+      ENTERING: 'entering',
+      LEAVING: 'leaving',
+    } as const;
+
+    /**
+     * Current state.
+     */
+    state: 'entering' | 'leaving' = null;
 
     /**
      * Get the transition target.
@@ -98,8 +147,11 @@ export function withTransition<S extends Base>(
      * Trigger the enter transition.
      */
     async enter(target?: HTMLElement | HTMLElement[]): Promise<void> {
+      const { STATES, EVENTS } = this.constructor;
       const { enterFrom, enterActive, enterTo, enterKeep, leaveTo } = this.$options;
-
+      this.state = STATES.ENTERING;
+      this.$emit(EVENTS.TRANSITION_ENTER);
+      this.$emit(EVENTS.TRANSITION_ENTER_START);
       removeClass(target ?? this.targets, leaveTo);
       await nextFrame();
       await transition(
@@ -111,14 +163,18 @@ export function withTransition<S extends Base>(
         },
         enterKeep ? 'keep' : undefined,
       );
+      this.$emit(EVENTS.TRANSITION_ENTER_END);
     }
 
     /**
      * Trigger the leave transition.
      */
     async leave(target?: HTMLElement | HTMLElement[]): Promise<void> {
+      const { STATES, EVENTS } = this.constructor;
       const { leaveFrom, leaveActive, leaveTo, leaveKeep, enterTo } = this.$options;
-
+      this.state = STATES.LEAVING;
+      this.$emit(EVENTS.TRANSITION_LEAVE);
+      this.$emit(EVENTS.TRANSITION_LEAVE_START);
       removeClass(target ?? this.targets, enterTo);
       await nextFrame();
       await transition(
@@ -130,6 +186,25 @@ export function withTransition<S extends Base>(
         },
         leaveKeep ? 'keep' : undefined,
       );
+      this.$emit(EVENTS.TRANSITION_LEAVE_END);
+    }
+
+    /**
+     * Toggle the leave or enter transition.
+     * Defaults to the enter transition if no transition has been triggered yet.
+     */
+    async toggle(target?: HTMLElement | HTMLElement[]): Promise<void> {
+      const { STATES, EVENTS } = this.constructor;
+
+      this.$emit(EVENTS.TRANSITION_TOGGLE);
+
+      switch (this.state) {
+        case STATES.ENTERING:
+          return this.leave(target);
+        case STATES.LEAVING:
+        default:
+          return this.enter(target);
+      }
     }
   }